Private Types

Wing Geometry, Panel and Aerodynamics

VortexStepMethod.Panel — Type

@with_kw mutable struct Panel

Represents a panel in a vortex step method simulation. All points and vectors are in the kite body (KB) frame.

Fields

TE_point_1::MVec3=zeros(MVec3): First trailing edge point

LE_point_1::MVec3=zeros(MVec3): First leading edge point

TE_point_2::MVec3=zeros(MVec3): Second trailing edge point

LE_point_2::MVec3=zeros(MVec3): Second leading edge point

chord::Float64=0: Panel chord length

va::MVec3=zeros(MVec3): Panel velocity

corner_points::MMatrix{3, 4, Float64}=zeros(MMatrix{3, 4, Float64}: Panel corner points

aero_model::AeroModel=INVISCID: Aerodynamic model type AeroModel

aero_center::Vector{Float64}: Panel aerodynamic center

cl_coeffs::Vector{Float64}=zeros(Float64, 3)

cd_coeffs::Vector{Float64}=zeros(Float64, 3)

cm_coeffs::Vector{Float64}=zeros(Float64, 3)

cl_interp::Union{Nothing, I1, I2} = nothing

cd_interp::Union{Nothing, I1, I2} = nothing

cm_interp::Union{Nothing, I1, I2} = nothing

control_point::Vector{MVec3}: Panel control point

bound_point_1::Vector{MVec3}: First bound point

bound_point_2::Vector{MVec3}: Second bound point

x_airf::MVec3=zeros(MVec3): Unit vector perpendicular to chord line

y_airf::MVec3=zeros(MVec3): Unit vector parallel to chord line

z_airf::MVec3=zeros(MVec3): Unit vector in spanwise direction

width::Float64=0: Panel width

filaments::Tuple{BoundFilament,BoundFilament,BoundFilament,SemiInfiniteFilament,SemiInfiniteFilament} = ( BoundFilament(), BoundFilament(), BoundFilament(), SemiInfiniteFilament(), SemiInfiniteFilament() ): Panel filaments, see: BoundFilament

VortexStepMethod.PanelProperties — Type

PanelProperties

Structure to hold calculated panel properties.

Fields

aero_centers::Vector{MVec3}

control_points::Vector{MVec3}

bound_points_1::Vector{MVec3}

bound_points_2::Vector{MVec3}

x_airf::Vector{MVec3}: Vector of unit vectors perpendicular to chord line

y_airf::Vector{MVec3}: Vector of unit vectors parallel to chord line

z_airf::Vector{MVec3}: Vector of unit vectors in spanwise direction

VortexStepMethod.BoundFilament — Type

BoundFilament

Represents a bound vortex filament defined by two points.

Fields

x1::MVec3=zeros(MVec3): First point

x2::MVec3=zeros(MVec3): Second point

length=zero(Float64): Filament length

r0::MVec3=zeros(MVec3): Vector from x1 to x2

initialized::Bool = false